+++ changelog.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented here.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [2.1.0]
 ### Added
 - `endDate` to `ProductDiscount`
 - `getBalance()` to `OrganisationService`
@@ -186,7 +186,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [< 0.3.1]
 ### History unavailable
 
-[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/2.0.2...HEAD
+[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/2.1.0...HEAD
+[2.1.0]: https://github.com/ordercloud/ordercloud-php/compare/2.0.2...2.1.0
 [2.0.2]: https://github.com/ordercloud/ordercloud-php/compare/2.0.1...2.0.2
 [2.0.1]: https://github.com/ordercloud/ordercloud-php/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/ordercloud/ordercloud-php/compare/1.6.0...2.0.0